# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  v1.ErrorResponse:
    properties:
      error:
        example: email already registered
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: solarwind.palm@gmail.com
    name: Alexander Cheryomukhin
  description: Grom API documentation server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Grom API
  version: "1.0"
paths:
...
  /users/{nickname}/avatar:
    get:
      description: Return avatar image for a local user
      parameters:
      - description: User nickname
        in: path
        name: nickname
        required: true
        type: string
      produces:
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user avatar
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"